Fix zero iteration bug
The bug behavior is that no media were being logged upon running the
puller. There were a few contributing factors. First, the media id
limit was too low, causing the program to terminate early. Second, our
rates for dropout were too high. When the dropouts nested, the rates
multiply, causing retained reviews to be far too rare.

We've also moved some of the work of the filters into the GraphQL
queries. This drastically decreases the number of polls we do with the
API, which helps reduce the request limit bottleneck.